bench: switch to criterion
