+++ Cargo.toml
@@ -108,6 +108,7 @@ rand_xorshift = "0.3"
 regex = "^1.7"
 serde = { version = "^1.0", features = ["derive"] }
 serde_json = "^1.0"
+criterion = { version = "0.4", features = ["html_reports"] }
 
 [features]
 default = []
@@ -118,3 +119,11 @@ wasm = ["chrono/wasmbind", "getrandom", "getrandom/js"]
 
 [profile.bench]
 debug = true
+
+[[bench]]
+name = "key"
+harness = false
+
+# [[bench]]
+# name = "message"
+# harness = false
